feat(web): Combined graph and table to page
